Add NotImplemented to functions so that code will compile and I can add definitions later

+++ my_libs/ode/src/solve_ode.cpp
@@ -16,7 +16,7 @@ Col<T> SolveODE(const function<Col<T>(Col<T>)>& func, const Col<T> x0, double dt
     bool use_explicit = true;
     for (auto e_i: eigen_values) {
         const double mod = std::sqrt(
-            e_i.imag * e_i.imag + e_i.real * e_i.real
+            (double)e_i.imag * (double)e_i.imag + (double)e_i.real * (double)e_i.real
         );
 
         const double z = mod * dt;
